Abstract API Key retrieval functionality into its own class
And use it in Etherscan and Cryptocompare
